+++ LibCore/Utils/Collections/Extensions/ListExtensions.cs
@@ -63,10 +63,7 @@ public static class ListExtensions
     /// </summary>
     public static void AddAll< T >( this List< T > target, T[] source, int start, int count )
     {
-        for ( var i = start; i < count; i++ )
-        {
-            target.Add( source[ i ] );
-        }
+        for ( var i = start; i < count; i++ ) target.Add( source[ i ] );
     }
 
     /// <summary>
@@ -75,10 +72,7 @@ public static class ListExtensions
     /// </summary>
     public static void AddAll< T >( this List< T > target, List< T > source, int start, int count )
     {
-        for ( var i = start; i < count; i++ )
-        {
-            target.Add( source[ i ] );
-        }
+        for ( var i = start; i < count; i++ ) target.Add( source[ i ] );
     }
 
     /// <summary>
@@ -86,10 +80,7 @@ public static class ListExtensions
     /// </summary>
     public static void AddAll< T >( this List< T > target, List< T > source )
     {
-        foreach ( T tex in source )
-        {
-            target.Add( tex );
-        }
+        foreach ( var tex in source ) target.Add( tex );
     }
 
     /// <summary>
@@ -97,10 +88,7 @@ public static class ListExtensions
     /// </summary>
     public static void AddAll< T >( this List< T > target, params T[] items )
     {
-        foreach ( T item in items )
-        {
-            target.Add( item );
-        }
+        foreach ( var item in items ) target.Add( item );
     }
 
     /// <summary>
@@ -127,10 +115,7 @@ public static class ListExtensions
     /// </summary>
     public static void Truncate< T >( this List< T > ts, int newSize )
     {
-        if ( ts.Count > newSize )
-        {
-            ts.RemoveRange( newSize, ts.Count - newSize );
-        }
+        if ( ts.Count > newSize ) ts.RemoveRange( newSize, ts.Count - newSize );
     }
 
     /// <summary>
@@ -142,12 +127,9 @@ public static class ListExtensions
     /// <exception cref="GdxRuntimeException"></exception>
     public static T Pop< T >( this List< T > list )
     {
-        if ( list.Count == 0 )
-        {
-            throw new GdxRuntimeException( "List is empty." );
-        }
+        if ( list.Count == 0 ) throw new GdxRuntimeException( "List is empty." );
 
-        T item = list[ ^1 ];
+        var item = list[ ^1 ];
 
         list.RemoveAt( list.Count - 1 );
 
@@ -174,12 +156,9 @@ public static class ListExtensions
     /// <returns></returns>
     public static T RemoveIndex< T >( this List< T > list, int index )
     {
-        if ( index >= list.Count )
-        {
-            throw new IndexOutOfRangeException( $"index can't be >= size: {index} >= {list.Count}" );
-        }
+        if ( index >= list.Count ) throw new IndexOutOfRangeException( $"index can't be >= size: {index} >= {list.Count}" );
 
-        T value = list[ index ];
+        var value = list[ index ];
 
         list.RemoveAt( index );
